finish tutorial

+++ app/Http/Controllers/ListingController.php
@@ -47,6 +47,7 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos','public');
         }
 
+        $formFields['user_id'] = auth()->id();
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing crated successfully!');
@@ -61,6 +62,12 @@ class ListingController extends Controller
     // Update listing data
     public function update(Request $request, Listing $listing){
         // dd($request->file('logo'));
+
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -82,8 +89,16 @@ class ListingController extends Controller
     
     // Delete listing
     public function destroy(Listing $listing){
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
     }
 
+    // Manage Listings
+    public function manage(){
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
+    }
 }
