+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\ListingController;
+use App\Http\Controllers\UserController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Models\Listing;
@@ -21,20 +22,38 @@ use PhpParser\Node\Expr\List_;
 Route::get('/', [ListingController::class, 'index']);
 
 // create - Show form to create new listing
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // store - Store new listing
 Route::post('/listings', [ListingController::class, 'store']);
 
 // edit - Show form to edit listing
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // update - Update listing
-Route::put('//listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // destroy - Delete listing  
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
+// Manage listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 //place in the last
 // show - Show single listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
+
+// show - Show register user
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+
+// create new user
+Route::post('/users', [UserController::class, 'store']);
+
+// Log user out
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+// Show login Form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+// Log in User
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
+
